use different resolutions for Camera and CNN-input
Before this commit, the camera resolution used to be the same as the
resolution of the image fed to the CNN as input (160x120 by default).

But there are many webcams which do not support 160x120.
Moreover, if the user changes the image dimensions used for the CNN,
it is very unlikely that a webcam will support that resolution.

Hence, from now on, the webcam will capture frames on the very commonly
supported resolution 320x240. This will be then downscaled to 160x120
for feeding into the CNN.